third_party/WebKit/LayoutTests/http/tests/workers/worker-importScripts-onerror-redirect-to-crossorigin-expected.txt

Ensure that scripts imported into a Worker from a redirect to a cross-origin host trigger sanitized onerror messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS workerOnerror.message is "Script error."
PASS workerOnerror.filename is ""
PASS workerOnerror.lineno is 0
PASS workerOnerror.colno is 0
PASS workerOnerror.error is null
PASS pageOnerror.message is "Script error."
PASS pageOnerror.filename is ""
PASS pageOnerror.lineno is 0
PASS pageOnerror.colno is 0
PASS pageOnerror.error is null
PASS successfullyParsed is true

TEST COMPLETE
